+++ packages/editor/editor-core/src/plugins/insert-block/ui/ToolbarInsertBlock/index.tsx
@@ -62,7 +62,6 @@ import { showLinkToolbar } from '../../../hyperlink/commands';
 import { insertMentionQuery } from '../../../mentions/commands/insert-mention-query';
 import { updateStatus } from '../../../status/actions';
 import {
-  AnalyticsEventPayload,
   withAnalytics as commandWithAnalytics,
   ACTION,
   ACTION_SUBJECT,
@@ -71,6 +70,7 @@ import {
   ACTION_SUBJECT_ID,
   PANEL_TYPE,
   InsertEventPayload,
+  DispatchAnalyticsEvent,
 } from '../../../analytics';
 import { EditorState } from 'prosemirror-state';
 
@@ -80,14 +80,29 @@ export const messages = defineMessages({
     defaultMessage: 'Action item',
     description: 'Also known as a “task”, “to do item”, or a checklist',
   },
+  actionDescription: {
+    id: 'fabric.editor.action.description',
+    defaultMessage: 'Capture actions to move work forward',
+    description: '',
+  },
   link: {
     id: 'fabric.editor.link',
     defaultMessage: 'Link',
     description: 'Insert a hyperlink',
   },
+  linkDescription: {
+    id: 'fabric.editor.link.description',
+    defaultMessage: 'Link to an internal or external page',
+    description: 'Insert a hyperlink',
+  },
   filesAndImages: {
     id: 'fabric.editor.filesAndImages',
     defaultMessage: 'Files & images',
+    description: 'Insert one or more files or images',
+  },
+  filesAndImagesDescription: {
+    id: 'fabric.editor.filesAndImages.description',
+    defaultMessage: 'Add images and other files to your page',
     description: 'Insert one or more files or images',
   },
   image: {
@@ -100,9 +115,19 @@ export const messages = defineMessages({
     defaultMessage: 'Mention',
     description: 'Reference another person in your document',
   },
+  mentionDescription: {
+    id: 'fabric.editor.mention.description',
+    defaultMessage: 'Mention someone to send them a notification',
+    description: 'Reference another person in your document',
+  },
   emoji: {
     id: 'fabric.editor.emoji',
     defaultMessage: 'Emoji',
+    description: 'Insert an emoticon or smiley :-)',
+  },
+  emojiDescription: {
+    id: 'fabric.editor.emoji.description',
+    defaultMessage: 'Use emojis to express ideas 🎉 and emotions 😄',
     description: 'Insert an emoticon or smiley :-)',
   },
   table: {
@@ -110,9 +135,19 @@ export const messages = defineMessages({
     defaultMessage: 'Table',
     description: 'Inserts a table in the document',
   },
+  tableDescription: {
+    id: 'fabric.editor.table.description',
+    defaultMessage: 'Insert a table',
+    description: 'Inserts a table in the document',
+  },
   decision: {
     id: 'fabric.editor.decision',
     defaultMessage: 'Decision',
+    description: 'Capture a decision you’ve made',
+  },
+  decisionDescription: {
+    id: 'fabric.editor.decision.description',
+    defaultMessage: 'Capture decisions so they’re easy to track',
     description: 'Capture a decision you’ve made',
   },
   horizontalRule: {
@@ -120,9 +155,19 @@ export const messages = defineMessages({
     defaultMessage: 'Divider',
     description: 'A horizontal rule or divider',
   },
+  horizontalRuleDescription: {
+    id: 'fabric.editor.horizontalRule.description',
+    defaultMessage: 'Separate content with a horizontal line',
+    description: 'A horizontal rule or divider',
+  },
   date: {
     id: 'fabric.editor.date',
     defaultMessage: 'Date',
+    description: 'Opens a date picker that lets you select a date',
+  },
+  dateDescription: {
+    id: 'fabric.editor.date.description',
+    defaultMessage: 'Add a date using a calendar',
     description: 'Opens a date picker that lets you select a date',
   },
   placeholderText: {
@@ -135,9 +180,20 @@ export const messages = defineMessages({
     defaultMessage: 'Columns',
     description: 'Create a multi column section or layout',
   },
+  columnsDescription: {
+    id: 'fabric.editor.columns.description',
+    defaultMessage: 'Structure your page using sections',
+    description: 'Create a multi column section or layout',
+  },
   status: {
     id: 'fabric.editor.status',
     defaultMessage: 'Status',
+    description:
+      'Inserts an item representing the status of an activity to task.',
+  },
+  statusDescription: {
+    id: 'fabric.editor.status.description',
+    defaultMessage: 'Create a colored lozenge with text inside',
     description:
       'Inserts an item representing the status of an activity to task.',
   },
@@ -195,7 +251,7 @@ export interface Props {
     node?: PMNode,
     isEditing?: boolean,
   ) => (state: EditorState, dispatch: CommandDispatch) => void;
-  dispatchAnalyticsEvent?: (payload: AnalyticsEventPayload) => void;
+  dispatchAnalyticsEvent?: DispatchAnalyticsEvent;
 }
 
 export interface State {