+++ packages/elements/emoji/src/api/media/SiteEmojiResource.ts
@@ -18,7 +18,7 @@ import {
   isMediaEmoji,
   convertImageToMediaRepresentation,
   isLoadedMediaEmoji,
-} from '../../type-helpers';
+} from '../../util/type-helpers';
 import MediaEmojiCache from './MediaEmojiCache';
 import {
   denormaliseEmojiServiceResponse,