+++ packages/elements/share/CHANGELOG.md
@@ -1,5 +1,10 @@
 # @atlaskit/share
 
+## 0.2.9
+- [patch] [fcdae04b8c](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/fcdae04b8c):
+
+  - FS-3620 add share analytics
+
 ## 0.2.8
 - [patch] [13d9986e40](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/13d9986e40):