+++ packages/navigation/atlassian-switcher/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @atlaskit/atlassian-switcher
 
+## 0.4.0
+- [minor] [e36f791fd6](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/e36f791fd6):
+
+  - Improve types
+
+## 0.3.6
+- [patch] [db2a7ffde6](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/db2a7ffde6):
+
+  - Fixing recent containers bug
+
 ## 0.3.5
 - [patch] [9d6f8d516a](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/9d6f8d516a):